requirements.txt
langchain>=0.2.9,<1.0
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-chroma
langchain-unstructured
python-dotenv
pydantic
duckduckgo-search
tavily-python
atlassian-python-api
gradio
pytest
chromadb
unstructured
jinja2
langgraph
importlib
logging